Final, working version
Whole RSA algorithm was moved inside main for loop. This way, if a picture has more than one IDAT chunk, algorithm encodes all of those chunks.
Program also creates two new files. One of them is a pictue with encrypted IDAT chunks and the other one has those chunks decrypted (to see if the output is the same as the input). Names of created files are: icon-po-encrypted and icon-po-decrypted.